requirements.txt
discord.py>=1.6
python-dotenv==0.15.0
discord-ext-typed-commands==1.0.3
